third_party/blink/web_tests/platform/generic/svg/css/mask-type-expected.txt

Test that clip-path shapes accept different length units

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle("mask-type", "") is "luminance"
PASS computedStyle("mask-type", "alpha") is "alpha"
PASS computedStyle("mask-type", "luminance") is "luminance"
PASS computedStyle("mask-type", "0") is "luminance"
PASS computedStyle("mask-type", "1") is "luminance"
PASS computedStyle("mask-type", "linearRGB") is "luminance"
PASS successfullyParsed is true

TEST COMPLETE
